requirements.txt
eze-cli
pydash